requirements.txt
requests==2.24.0
urllib3==1.25.10
instabot==0.117.0
Pillow==7.2.0
python-dotenv==0.14.0
